% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{dimsize}
\alias{dimsize}
\title{Helper function to get dimension sizes of a ncvar4 object.}
\usage{
dimsize(var, dimlist)
}
\arguments{
\item{var}{ncvar4 object}

\item{dimlist}{as returned by \code{ncss_dimlist()}}
}
\description{
Helper function to get dimension sizes of a ncvar4 object.
}
