% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getdim}
\alias{getdim}
\title{helper function to get ncdim4 object from nc by name}
\usage{
getdim(nc, dimname)
}
\arguments{
\item{nc}{a ncdf4 object as returned by \code{ncdf4::nc_open()}}

\item{dimname}{Name of ncdim4 object from \code{nc$dim}}
}
\description{
helper function to get ncdim4 object from nc by name
}
