% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{index_array}
\alias{index_array}
\title{Index an arbitrarily-sized array, from Hadley}
\usage{
index_array(x, dim, value, drop = FALSE)
}
\arguments{
\item{x}{an array}

\item{dim}{dimension along which to index}

\item{value}{values (vector) of indices}

\item{drop}{passed to \code{[}}
}
\description{
Copied from https://stackoverflow.com/a/14502298.
}
