% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{nc_subset}
\alias{nc_subset}
\title{Subset a netcdf}
\usage{
nc_subset(nc, ..., filename = tempfile(), keep_open = TRUE)
}
\arguments{
\item{nc}{a ncdf4 object, as returned by \code{ncdf4::nc_open()}}

\item{...}{\code{dplyr::filter}-like expressions. See details.}

\item{filename}{Where to write the new netcdf? Defaults to \code{tempfile()}.}

\item{keep_open}{Keep the netcdf open? Defaults to TRUE.}
}
\description{
Creates (and writes) a subset of a ncdf4 object using dplyr-like expressions
}
\details{
The new netcdf is a subset of the original, and also includes new 1-D
 variables containing the original dimensions' values.
}
