% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{ncss_create_fill}
\alias{ncss_create_fill}
\title{Create and fill a new netcdf as a subset of an existing netcdf}
\usage{
ncss_create_fill(nc, filename, varlist, indlist, keep_open = TRUE,
  optimize = TRUE)
}
\arguments{
\item{nc}{a ncdf4 object, as returned by \code{ncdf4::nc_open()}}

\item{filename}{Where to write the new netcdf? Defaults to \code{tempfile()}.}

\item{varlist}{as returned by \code{ncss_varlist()}}

\item{indlist}{as returned by \code{ncss_indlist()}}

\item{keep_open}{Keep the netcdf open? Defaults to TRUE.}
}
\description{
Used internally by \code{nc_subset()}
}
\details{
The new netcdf is a subset of the original, and also includes new 1-D
 variables containing the original dimensions' values.
}
