% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{ncss_dimlist}
\alias{ncss_dimlist}
\title{Create a list of ncdimr4 objects from a subset list}
\usage{
ncss_dimlist(nc, indlist)
}
\arguments{
\item{nc}{ncdf4 object, as returned by \code{ncdf4::nc_open()}}

\item{indlist}{As returne by \code{ncss_indlist()}}
}
\description{
Create a list of ncdimr4 objects from a subset list
}
