% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{ncss_indlist}
\alias{ncss_indlist}
\title{Function to subset and write netcdf v4 files, including ones that have groups.}
\usage{
ncss_indlist(nc, ..., env = parent.frame())
}
\arguments{
\item{nc}{a ncdf4 object as returned by \code{ncdf4::nc_open()}}

\item{...}{Logical expressions involving variables in nc}

\item{env}{passed to \code{rlang::eval_tidy()}}
}
\description{
Returns a list that can be used with \code{ncdf4::nc_create()}
Return vector of indices (not dimension values)
}
