% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{ncss_varlist}
\alias{ncss_varlist}
\title{Create a list of ncvar4 objects from a subset list}
\usage{
ncss_varlist(nc, dimlist)
}
\arguments{
\item{nc}{ncdf4 object, as returned by \code{ncdf4::nc_open()}}

\item{dimlist}{As returne by \code{ncss_dimlist()}}
}
\description{
Create a list of ncvar4 objects from a subset list
}
