% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{ncvar_getss}
\alias{ncvar_getss}
\title{Get a subset of a single netcdf variable}
\usage{
ncvar_getss(nc, varid, indlist = NULL, verbose = FALSE,
  signedbyte = TRUE, raw_datavals = FALSE, optimize = TRUE)
}
\arguments{
\item{nc}{An object of class \code{ncdf4} (as returned by either 
 function \code{\link[ncdf4]{nc_open}}
 or function \code{\link[ncdf4]{nc_create}}), indicating what file to read from.}

\item{varid}{What variable to read the data from.  Can be a string with the name
 of the variable or an object of class \code{ncvar4}
 If left unspecified, the function will determine if there
 is only one variable in the file and, if so, read from that.  If left unspecified
 and there are multiple variables in the file, an error is generated.
 This argument can also, optionally, specify the name of a dimension (usually the
 unlimited dimension) in order to read values from a coordinate variable.
 Note this is not usual practice, because the \code{ncdim} object already
 contains all the dimension's values in the field named "vals".  However, it can
 sometimes be faster to turn off this automatic reading of the unlimited dimension's
 values by using \code{nc_open(filename, readunlim=FALSE)}, then read the
 dimension values in later with this function.}

\item{indlist}{As returned by \code{ncss_indlist()}}

\item{verbose}{If TRUE, then progress information is printed.}

\item{signedbyte}{If TRUE (default), then on-disk byte variables are interpreted
 as signed.  This is in accord with the netCDF standard.  If FALSE, then
 on-disk byte variables are interpreted as unsigned.}

\item{raw_datavals}{If TRUE, then the actual raw data values from the
 file are returned with no conversion to NA (if equal to the missing value/fill value) or
 scale/offset applied. Default is FALSE.}

\item{optimize}{optimize call to \code{ncdf4::ncvar_get}}
}
\description{
Note: \code{collapse_degen} (from \code{ncdf4::ncvar_get}
 is always FALSE here, in order to
 allow subsetting of resulting arrays.
}
