% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{reassign_bytes}
\alias{reassign_bytes}
\title{Helper function to workaround ncdf4's prec limitations}
\usage{
reassign_bytes(varlist)
}
\description{
Helper function to workaround ncdf4's prec limitations
}
