% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reassign_prec}
\alias{reassign_prec}
\title{Helper function to workaround ncdf4's prec limitations}
\usage{
reassign_prec(varlist)
}
\arguments{
\item{varlist}{as returned by \code{nc_varlist()}}
}
\description{
Helper function to workaround ncdf4's prec limitations
}
