% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ssdim}
\alias{ssdim}
\title{Helper function to subset a ncdim4 object}
\usage{
ssdim(ncdim, keepinds)
}
\arguments{
\item{ncdim}{a ncdim4 object}

\item{keepinds}{indices to keep}
}
\description{
Helper function to subset a ncdim4 object
}
