% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{getdim}
\alias{getdim}
\title{Function to subset and write netcdf v4 files, including ones that have groups.}
\usage{
getdim(nc, dimname)
}
\arguments{
\item{nc}{a ncdf4 object as returned by \code{ncdf4::nc_open()}}

\item{dimname}{Name of ncdim4 object from \code{nc$dim}}
}
\description{
Returns a list that can be used with \code{ncdf4::nc_create()}
helper function to get ncdim4 object from nc by name
}
