% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{getExprNames}
\alias{getExprNames}
\title{Helper function to recursively get names from an expression}
\usage{
getExprNames(expr)
}
\description{
Helper function to recursively get names from an expression
}
