% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{getExprNames}
\alias{getExprNames}
\title{Function to subset and write netcdf v4 files, including ones that have groups.}
\usage{
getExprNames(expr)
}
\description{
Returns a list that can be used with \code{ncdf4::nc_create()}
Helper function to recursively get names from an expression
}
