% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_subset.R
\name{ncss_indlist}
\alias{ncss_indlist}
\title{Return vector of indices (not dimension values)}
\usage{
ncss_indlist(nc, ...)
}
\arguments{
\item{nc}{a ncdf4 object as returned by \code{ncdf4::nc_open()}}

\item{...}{Logical expressions involving variables in nc}
}
\description{
Return vector of indices (not dimension values)
}
